requirements.txt
requests>=2.28.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pystray>=0.19.0
Pillow>=10.0.0
